tailor_message_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
litellm==1.30.7
python-dotenv==1.0.1
arize==7.0.1
openinference-instrumentation-litellm==0.0.2
pydantic==2.6.1 